fix edit profile
