+++ back/backend/handlers/admin/show_users.py
@@ -1,15 +1,12 @@
 from json import loads
-import datetime
 from flask import Blueprint, jsonify, request
 from google.cloud import ndb
 
 from backend.models.wca.person import Person
-from backend.models.wca.rank import RankAverage, RankSingle
-from backend.lib import permissions, auth
-from backend.models.province import Province
-from backend.models.user import User, UserLocationUpdate, Roles
+from backend.lib import auth
+from backend.models.user import User, Roles
 
-bp = Blueprint('edit_users', __name__)
+bp = Blueprint('show_users', __name__)
 client = ndb.Client()
 
 
@@ -102,74 +99,3 @@ def get_users_by_id():
             {'data': data}
         )
 
-
-# After updating the user's province, write the RankSingle and RankAverage to the
-# datastore again to update their provinces.
-def RewriteRanks(wca_person):
-    if not wca_person:
-        return
-    for rank_class in (RankSingle, RankAverage):
-        ndb.put_multi(rank_class.query(rank_class.person == wca_person.key).fetch())
-
-
-@bp.route('/edit', methods=['POST'])
-@bp.route('/edit/<user_id>', methods=['POST'])
-def edit(user_id=-1):
-    with client.context():
-        me = auth.user()
-        if not me:
-            return jsonify({"error": "Unauthorized"}), 401
-        if user_id == -1:
-            user = me
-        else:
-            user = User.get_by_id(user_id)
-        if not user:
-            return jsonify({"error": "Unrecognized user ID %s" % user_id}), 404
-        if not permissions.CanViewUser(user, me):
-            return jsonify(
-                {"error": "You're not authorized to view this user. So you can't edit their location either."}), 403
-
-        province_id = request.json['province']
-        if province_id == 'na':
-            province_id = ''
-
-        old_province_id = user.province.id() if user.province else ''
-        changed_location = old_province_id != province_id
-        user_modified = False
-        if permissions.CanEditLocation(user, me) and changed_location:
-            if province_id:
-                user.province = ndb.Key(Province, province_id)
-            else:
-                del user.province
-            if user.wca_person and old_province_id != province_id:
-                wca_person = user.wca_person.get()
-                if wca_person:
-                    wca_person.province = user.province
-                    wca_person.put()
-                RewriteRanks(wca_person)
-            user_modified = True
-
-            if changed_location:
-                # Also save the Update.
-                update = UserLocationUpdate()
-                update.updater = me.key
-                update.update_time = datetime.datetime.now()
-                if province_id:
-                    update.province = ndb.Key(Province, province_id)
-                user.updates.append(update)
-
-        elif changed_location:
-            return jsonify({"error": "You're not authorized to edit this user's location."}), 403
-
-        for role in permissions.EditableRoles(user, me):
-            if role in request.json["roles"] and role not in user.roles:
-                user.roles.append(role)
-                user_modified = True
-            elif role not in request.json["roles"] and role in user.roles:
-                user.roles.remove(role)
-                user_modified = True
-
-        if user_modified:
-            user.put()
-
-        return user.toJson()
